LayoutTests/media/modern-media-controls/media-controller/ios/media-controller-ios-only-enable-tap-gesture-recognizer-with-fades-when-idle-expected.txt

Testing the MediaController controls bar tap gesture recognizer is only enabled when fadesWhileIdle is true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default state
PASS autoHideController.fadesWhileIdle is false
PASS autoHideController._tapGestureRecognizer.enabled is false

autoHideController.fadesWhileIdle = true
PASS autoHideController.fadesWhileIdle is true
PASS autoHideController._tapGestureRecognizer.enabled is true

autoHideController.fadesWhileIdle = false
PASS autoHideController.fadesWhileIdle is false
PASS autoHideController._tapGestureRecognizer.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
